core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "utcp"
version = "1.0.4"
authors = [
  { name = "UTCP Contributors" },
]
description = "Universal Tool Calling Protocol (UTCP) client library for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tomli>=2.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MPL-2.0"

[project.optional-dependencies]
dev = [
    "build",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "coverage",
    "twine",
]

[project.urls]
Homepage = "https://utcp.io"
Source = "https://github.com/universal-tool-calling-protocol/python-utcp"
Issues = "https://github.com/universal-tool-calling-protocol/python-utcp/issues"